Implements Parts of #75
Broke down the cli helper file
